# ENH: spatial: enable non-double dtypes in squareform