--- a/scipy/spatial/distance.py
+++ b/scipy/spatial/distance.py
@@ -124,26 +124,15 @@ def _copy_array_if_base_present(a):
     """
     if a.base is not None:
         return a.copy()
-    elif np.issubsctype(a, np.float32):
-        return np.array(a, dtype=np.double)
-    else:
-        return a
+    return a
 
 
 def _convert_to_bool(X):
-    if X.dtype != bool:
-        X = X.astype(bool)
-    if not X.flags.contiguous:
-        X = X.copy()
-    return X
+    return np.ascontiguousarray(X, dtype=bool)
 
 
 def _convert_to_double(X):
-    if X.dtype != np.double:
-        X = X.astype(np.double)
-    if not X.flags.contiguous:
-        X = X.copy()
-    return X
+    return np.ascontiguousarray(X, dtype=np.double)
 
 
 def _validate_vector(u, dtype=None):
@@ -1447,7 +1436,7 @@ def squareform(X, force="no", checks=True):
 
     """
 
-    X = _convert_to_double(np.asarray(X, order='c'))
+    X = np.ascontiguousarray(X)
 
     s = X.shape
 
@@ -1462,21 +1451,21 @@ def squareform(X, force="no", checks=True):
 
     # X = squareform(v)
     if len(s) == 1:
-        if X.shape[0] == 0:
-            return np.zeros((1, 1), dtype=np.double)
+        if s[0] == 0:
+            return np.zeros((1, 1), dtype=X.dtype)
 
         # Grab the closest value to the square root of the number
         # of elements times 2 to see if the number of elements
         # is indeed a binomial coefficient.
-        d = int(np.ceil(np.sqrt(X.shape[0] * 2)))
+        d = int(np.ceil(np.sqrt(s[0] * 2)))
 
         # Check that v is of valid dimensions.
-        if d * (d - 1) / 2 != int(s[0]):
+        if d * (d - 1) != s[0] * 2:
             raise ValueError('Incompatible vector size. It must be a binomial '
                              'coefficient n choose 2 for some integer n >= 2.')
 
         # Allocate memory for the distance matrix.
-        M = np.zeros((d, d), dtype=np.double)
+        M = np.zeros((d, d), dtype=X.dtype)
 
         # Since the C code does not support striding using strides.
         # The dimensions are used instead.
@@ -1497,10 +1486,10 @@ def squareform(X, force="no", checks=True):
         d = s[0]
 
         if d <= 1:
-            return np.array([], dtype=np.double)
+            return np.array([], dtype=X.dtype)
 
         # Create a vector.
-        v = np.zeros((d * (d - 1)) // 2, dtype=np.double)
+        v = np.zeros((d * (d - 1)) // 2, dtype=X.dtype)
 
         # Since the C code does not support striding using strides.
         # The dimensions are used instead.
@@ -1511,16 +1500,16 @@ def squareform(X, force="no", checks=True):
         return v
     else:
         raise ValueError(('The first argument must be one or two dimensional '
-                         'array. A %d-dimensional array is not '
-                         'permitted') % len(s))
+                          'array. A %d-dimensional array is not '
+                          'permitted') % len(s))
 
 
 def is_valid_dm(D, tol=0.0, throw=False, name="D", warning=False):
     """
     Returns True if input array is a valid distance matrix.
 
-    Distance matrices must be 2-dimensional numpy arrays containing
-    doubles. They must have a zero-diagonal, and they must be symmetric.
+    Distance matrices must be 2-dimensional numpy arrays.
+    They must have a zero-diagonal, and they must be symmetric.
 
     Parameters
     ----------
@@ -1556,13 +1545,6 @@ def is_valid_dm(D, tol=0.0, throw=False, name="D", warning=False):
     valid = True
     try:
         s = D.shape
-        if D.dtype != np.double:
-            if name:
-                raise TypeError(('Distance matrix \'%s\' must contain doubles '
-                                 '(double).') % name)
-            else:
-                raise TypeError('Distance matrix must contain doubles '
-                                '(double).')
         if len(D.shape) != 2:
             if name:
                 raise ValueError(('Distance matrix \'%s\' must have shape=2 '
@@ -1614,9 +1596,9 @@ def is_valid_y(y, warning=False, throw=False, name=None):
     """
     Returns True if the input array is a valid condensed distance matrix.
 
-    Condensed distance matrices must be 1-dimensional
-    numpy arrays containing doubles. Their length must be a binomial
-    coefficient :math:`{n \\choose 2}` for some positive integer n.
+    Condensed distance matrices must be 1-dimensional numpy arrays.
+    Their length must be a binomial coefficient :math:`{n \\choose 2}`
+    for some positive integer n.
 
     Parameters
     ----------
@@ -1638,13 +1620,6 @@ def is_valid_y(y, warning=False, throw=False, name=None):
     y = np.asarray(y, order='c')
     valid = True
     try:
-        if y.dtype != np.double:
-            if name:
-                raise TypeError(('Condensed distance matrix \'%s\' must '
-                                 'contain doubles (double).') % name)
-            else:
-                raise TypeError('Condensed distance matrix must contain '
-                                'doubles (double).')
         if len(y.shape) != 1:
             if name:
                 raise ValueError(('Condensed distance matrix \'%s\' must '