--- a/scipy/spatial/tests/test_distance.py
+++ b/scipy/spatial/tests/test_distance.py
@@ -1213,50 +1213,50 @@ class TestSomeDistanceFunctions(TestCase):
             assert_almost_equal(dist, np.sqrt(6.0))
 
 
-class TestSquareForm(TestCase):
+class TestSquareForm(object):
+    checked_dtypes = [np.float64, np.float32, np.int32, np.int8, bool]
 
-    def test_squareform_empty_matrix(self):
-        A = np.zeros((0,0))
-        rA = squareform(np.array(A, dtype='double'))
+    def test_squareform_matrix(self):
+        for dtype in self.checked_dtypes:
+            yield self.check_squareform_matrix, dtype
+
+    def test_squareform_vector(self):
+        for dtype in self.checked_dtypes:
+            yield self.check_squareform_vector, dtype
+
+    def check_squareform_matrix(self, dtype):
+        A = np.zeros((0,0), dtype=dtype)
+        rA = squareform(A)
         assert_equal(rA.shape, (0,))
+        assert_equal(rA.dtype, dtype)
 
-    def test_squareform_empty_vector(self):
-        v = np.zeros((0,))
-        rv = squareform(np.array(v, dtype='double'))
+        A = np.zeros((1,1), dtype=dtype)
+        rA = squareform(A)
+        assert_equal(rA.shape, (0,))
+        assert_equal(rA.dtype, dtype)
+
+        A = np.array([[0,4.2],[4.2,0]], dtype=dtype)
+        rA = squareform(A)
+        assert_equal(rA.shape, (1,))
+        assert_equal(rA.dtype, dtype)
+        assert_array_equal(rA, np.array([4.2], dtype=dtype))
+
+    def check_squareform_vector(self, dtype):
+        v = np.zeros((0,), dtype=dtype)
+        rv = squareform(v)
         assert_equal(rv.shape, (1,1))
-        assert_equal(rv[0, 0], 0)
+        assert_equal(rv.dtype, dtype)
+        assert_array_equal(rv, [[0]])
 
-    def test_squareform_1by1_matrix(self):
-        A = np.zeros((1,1))
-        rA = squareform(np.array(A, dtype='double'))
-        assert_equal(rA.shape, (0,))
-
-    def test_squareform_one_vector(self):
-        v = np.ones((1,)) * 8.3
-        rv = squareform(np.array(v, dtype='double'))
-        assert_equal(rv.shape, (2,2))
-        assert_equal(rv[0,1], 8.3)
-        assert_equal(rv[1,0], 8.3)
-
-    def test_squareform_one_binary_vector(self):
-        # Tests squareform on a 1x1 binary matrix; conversion to double was
-        # causing problems (see pull request 73).
-        v = np.ones((1,), dtype=bool)
+        v = np.array([8.3], dtype=dtype)
         rv = squareform(v)
         assert_equal(rv.shape, (2,2))
-        assert_(rv[0,1])
-
-    def test_squareform_2by2_matrix(self):
-        A = np.zeros((2,2))
-        A[0,1] = 0.8
-        A[1,0] = 0.8
-        rA = squareform(np.array(A, dtype='double'))
-        assert_equal(rA.shape, (1,))
-        assert_equal(rA[0], 0.8)
+        assert_equal(rv.dtype, dtype)
+        assert_array_equal(rv, np.array([[0,8.3],[8.3,0]], dtype=dtype))
 
     def test_squareform_multi_matrix(self):
         for n in xrange(2, 5):
-            yield self.check_squareform_multi_matrix(n)
+            yield self.check_squareform_multi_matrix, n
 
     def check_squareform_multi_matrix(self, n):
         X = np.random.rand(n, 4)
@@ -1373,15 +1373,6 @@ def is_valid_dm_throw(D):
 
 class TestIsValidDM(TestCase):
 
-    def test_is_valid_dm_int16_array_E(self):
-        # Tests is_valid_dm(*) on an int16 array. Exception expected.
-        D = np.zeros((5, 5), dtype='i')
-        assert_raises(TypeError, is_valid_dm_throw, (D))
-
-    def test_is_valid_dm_int16_array_F(self):
-        D = np.zeros((5, 5), dtype='i')
-        assert_equal(is_valid_dm(D), False)
-
     def test_is_valid_dm_improper_shape_1D_E(self):
         D = np.zeros((5,), dtype=np.double)
         assert_raises(ValueError, is_valid_dm_throw, (D))
@@ -1457,14 +1448,6 @@ class TestIsValidY(TestCase):
     # If test case name ends on "_E" then an exception is expected for the
     # given input, if it ends in "_F" then False is expected for the is_valid_y
     # check.  Otherwise the input is expected to be valid.
-
-    def test_is_valid_y_int16_array_E(self):
-        y = np.zeros((10,), dtype='i')
-        assert_raises(TypeError, is_valid_y_throw, (y))
-
-    def test_is_valid_y_int16_array_F(self):
-        y = np.zeros((10,), dtype='i')
-        assert_equal(is_valid_y(y), False)
 
     def test_is_valid_y_improper_shape_2D_E(self):
         y = np.zeros((3,3,), dtype=np.double)