This change makes the underlying C code for `scipy.spatial.distance.squareform` dtype-agnostic, and removes the checks and casts which previously would force all inputs to a double dtype.
Fixes gh-6433.

Informal timings show that (for double dtypes) expanding to the redundant matrix representation got about 2x slower, while condensing to the vector representation stayed about the same:

**this PR**
```python
v = np.random.random(4950)
m = squareform(v)
%timeit squareform(v)  # 10000 loops, best of 3: 62.9 µs per loop
%timeit squareform(m)  # 10000 loops, best of 3: 97.3 µs per loop
```
**current master**
```python
%timeit squareform(v)  # 10000 loops, best of 3: 32.6 µs per loop
%timeit squareform(m)  # 10000 loops, best of 3: 98.5 µs per loop
```

It's possible that there's a better way to write the vector->matrix transformation with `memcpy`, similar to what I did for the reverse operation, that will close this gap.

Finally, there is one back-compat concern: I removed the dtype checks from `is_valid_dm` and `is_valid_y`. I doubt that many users were relying on that particular case, but it's definitely a change to a public API.