@perimosocordiae 2016-08-04
With the second commit, the timings (using the same test setup) are now:
```python
%timeit squareform(v)  # 10000 loops, best of 3: 44.5 µs per loop
%timeit squareform(m)  # 10000 loops, best of 3: 95.2 µs per loop
```

So the slow case is now only ~1.5x slower than the previous behavior. To get the last bit of performance back, it might be necessary to write a specialized version for the common (double dtype) case.