@rgommers 2016-08-13
@fasiha this PR implements your enhancement request. Do you want to review it (would be helpful)?