@rgommers 2016-08-13
The 1.5x performance hit may be undesired and doesn't seem to be too hard to avoid with a specialized function. @juliantaylor spent some time last year on speeding up and releasing the GIL for this function, so that's at least one person who will likely care.