@perimosocordiae 2016-08-15
The latest commit adds a specialized version of the vector->matrix conversion for double-sized dtypes. The timings for double dtypes are now:

    %timeit squareform(v)  # 10000 loops, best of 3: 20.6 µs per loop
    %timeit squareform(m)  # 10000 loops, best of 3: 88.4 µs per loop

Looks like we're faster than current master for all cases, now.