@fasiha 2016-08-28
Wow, thanks @perimosocordiae! I (very belatedly…) looked through your changes per @rgommers’ message and will admit I am totally out of my league 😝, there’re so many t’s to cross and i’s to dot—it looks like you’ve covered all the bases but I can’t offer any more meaty feedback. Sorry, but please accept my thanks 🙇!