@perimosocordiae 2016-09-01
This PR is ready to merge: commits are squashed, and both v->m and m->v conversions are faster than current master for all dtypes.

Any further comments/concerns?