@rgommers 2016-09-18
Looks good to me now, thanks @perimosocordiae. Merging