@rgommers 2016-09-18
forgot about mentioning this change in release notes and docstring. follow-up in gh-6594